+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Install `prop-types` package in `dependencies` as official recommended.
 - Add new search and CRM icons.
 - Add new `<SearchInput>`.
+- Prefix component class names with `prefixClass()` helper.
 
 ### Changed
 - Upgrade `react` & `react-dom` to v15.5.